requirements.txt
streamlit==1.27.0 
pytesseract==0.3.10 
pdfplumber==0.9.0 
python-docx==0.8.11 
transformers==4.34.0 
sentence-transformers==2.2.2 
langchain==0.0.327